Clean up text label / text util
