+++ packages/tldraw/src/state/shapes/RectangleUtil/RectangleUtil.tsx
@@ -1,7 +1,5 @@
 import * as React from 'react'
 import { Utils, SVGContainer } from '@tldraw/core'
-import { Vec } from '@tldraw/vec'
-import { getStroke, getStrokePoints } from 'perfect-freehand'
 import { RectangleShape, DashStyle, TDShapeType, TDMeta } from '~types'
 import { GHOSTED_OPACITY } from '~constants'
 import { TDShapeUtil } from '../TDShapeUtil'
@@ -13,8 +11,11 @@ import {
   getFontStyle,
   transformSingleRectangle,
 } from '~state/shapes/shared'
-import { styled } from '~styles'
 import { TextLabel } from '../shared/TextLabel'
+import { getRectangleIndicatorPathTDSnapshot } from './helpers'
+import { DrawRectangle } from './components/DrawRectangle'
+import { DashedRectangle } from './components/DashedRectangle'
+import { BindingIndicator } from './components/BindingIndicator'
 
 type T = RectangleShape
 type E = SVGSVGElement
@@ -62,10 +63,9 @@ export class RectangleUtil extends TDShapeUtil<T, E> {
       ref
     ) => {
       const { id, size, style, text } = shape
-
+      const font = getFontStyle(shape.style)
       const styles = getShapeStyle(style, meta.isDarkMode)
-
-      const { strokeWidth } = styles
+      const Component = style.dash === DashStyle.Draw ? DrawRectangle : DashedRectangle
 
       const handleTextChange = React.useCallback(
         (text: string) => {
@@ -74,144 +74,38 @@ export class RectangleUtil extends TDShapeUtil<T, E> {
         [onShapeChange]
       )
 
-      const handleBlur = React.useCallback(() => {
-        onShapeBlur?.()
-      }, [onShapeBlur])
-
-      const font = getFontStyle(shape.style)
-
-      if (style.dash === DashStyle.Draw) {
-        const pathTDSnapshot = getRectanglePath(shape)
-        const indicatorPath = getRectangleIndicatorPathTDSnapshot(shape)
-
-        return (
-          <Wrapper>
-            <TextLabel
-              isEditing={isEditing}
-              onChange={handleTextChange}
-              onBlur={handleBlur}
-              text={text}
-              color={styles.stroke}
-              font={font}
-            />
-            <SVGContainer ref={ref} id={shape.id + '_svg'} {...events}>
-              {isBinding && (
-                <rect
-                  className="tl-binding-indicator"
-                  x={strokeWidth}
-                  y={strokeWidth}
-                  width={Math.max(0, size[0] - strokeWidth / 2)}
-                  height={Math.max(0, size[1] - strokeWidth / 2)}
-                  strokeWidth={this.bindingDistance * 2}
-                />
-              )}
-              <path
-                className={style.isFilled || isSelected ? 'tl-fill-hitarea' : 'tl-stroke-hitarea'}
-                d={indicatorPath}
-              />
-              <path
-                d={indicatorPath}
-                fill={style.isFilled ? styles.fill : 'none'}
-                pointerEvents="none"
-              />
-              <path
-                d={pathTDSnapshot}
-                fill={styles.stroke}
-                stroke={styles.stroke}
-                strokeWidth={styles.strokeWidth}
-                pointerEvents="none"
-                opacity={isGhost ? GHOSTED_OPACITY : 1}
-              />
-            </SVGContainer>
-          </Wrapper>
-        )
-      }
-
-      const sw = 1 + strokeWidth * 1.618
-
-      const w = Math.max(0, size[0] - sw / 2)
-      const h = Math.max(0, size[1] - sw / 2)
-
-      const strokes: [number[], number[], number][] = [
-        [[sw / 2, sw / 2], [w, sw / 2], w - sw / 2],
-        [[w, sw / 2], [w, h], h - sw / 2],
-        [[w, h], [sw / 2, h], w - sw / 2],
-        [[sw / 2, h], [sw / 2, sw / 2], h - sw / 2],
-      ]
-
-      const paths = strokes.map(([start, end, length], i) => {
-        const { strokeDasharray, strokeDashoffset } = Utils.getPerfectDashProps(
-          length,
-          strokeWidth * 1.618,
-          shape.style.dash
-        )
-
-        return (
-          <line
-            key={id + '_' + i}
-            x1={start[0]}
-            y1={start[1]}
-            x2={end[0]}
-            y2={end[1]}
-            strokeDasharray={strokeDasharray}
-            strokeDashoffset={strokeDashoffset}
-          />
-        )
-      })
-
       return (
-        <Wrapper>
-          <SVGContainer ref={ref} id={shape.id + '_svg'} {...events}>
-            <g opacity={isGhost ? GHOSTED_OPACITY : 1}>
-              {isBinding && (
-                <rect
-                  className="tl-binding-indicator"
-                  x={0}
-                  y={0}
-                  width={size[0]}
-                  height={size[1]}
-                  strokeWidth={this.bindingDistance}
-                />
-              )}
-              <rect
-                className={isSelected || style.isFilled ? 'tl-fill-hitarea' : 'tl-stroke-hitarea'}
-                x={sw / 2}
-                y={sw / 2}
-                width={w}
-                height={h}
-                strokeWidth={this.bindingDistance}
-              />
-              {style.isFilled && (
-                <rect
-                  x={sw / 2}
-                  y={sw / 2}
-                  width={w}
-                  height={h}
-                  fill={styles.fill}
-                  pointerEvents="none"
-                />
-              )}
-              <g pointerEvents="none" stroke={styles.stroke} strokeWidth={sw} strokeLinecap="round">
-                {paths}
-              </g>
-            </g>
+        <>
+          <TextLabel
+            isEditing={isEditing}
+            onChange={handleTextChange}
+            onBlur={onShapeBlur}
+            color={styles.stroke}
+            font={font}
+            text={text}
+          />
+          <SVGContainer
+            ref={ref}
+            id={shape.id + '_svg'}
+            opacity={isGhost ? GHOSTED_OPACITY : 1}
+            {...events}
+          >
+            {isBinding && <BindingIndicator strokeWidth={styles.strokeWidth} size={size} />}
+            <Component id={id} style={style} isSelected={isSelected} size={size} />
           </SVGContainer>
-        </Wrapper>
+        </>
       )
     }
   )
 
   Indicator = TDShapeUtil.Indicator<T>(({ shape }) => {
-    const {
-      style,
-      size: [width, height],
-    } = shape
+    const { id, style, size } = shape
 
     const styles = getShapeStyle(style, false)
     const sw = styles.strokeWidth
 
     if (style.dash === DashStyle.Draw) {
-      return <path d={getRectangleIndicatorPathTDSnapshot(shape)} />
+      return <path d={getRectangleIndicatorPathTDSnapshot(id, style, size)} />
     }
 
     return (
@@ -220,8 +114,8 @@ export class RectangleUtil extends TDShapeUtil<T, E> {
         y={sw}
         rx={1}
         ry={1}
-        width={Math.max(1, width - sw * 2)}
-        height={Math.max(1, height - sw * 2)}
+        width={Math.max(1, size[0] - sw * 2)}
+        height={Math.max(1, size[1] - sw * 2)}
       />
     )
   })
@@ -238,108 +132,3 @@ export class RectangleUtil extends TDShapeUtil<T, E> {
 
   transformSingle = transformSingleRectangle
 }
-
-/* -------------------------------------------------- */
-/*                       Helpers                      */
-/* -------------------------------------------------- */
-
-function getRectangleDrawPoints(shape: RectangleShape) {
-  const styles = getShapeStyle(shape.style)
-
-  const getRandom = Utils.rng(shape.id)
-
-  const sw = styles.strokeWidth
-
-  // Dimensions
-  const w = Math.max(0, shape.size[0])
-  const h = Math.max(0, shape.size[1])
-
-  // Random corner offsets
-  const offsets = Array.from(Array(4)).map(() => {
-    return [getRandom() * sw * 0.75, getRandom() * sw * 0.75]
-  })
-
-  // Corners
-  const tl = Vec.add([sw / 2, sw / 2], offsets[0])
-  const tr = Vec.add([w - sw / 2, sw / 2], offsets[1])
-  const br = Vec.add([w - sw / 2, h - sw / 2], offsets[2])
-  const bl = Vec.add([sw / 2, h - sw / 2], offsets[3])
-
-  // Which side to start drawing first
-  const rm = Math.round(Math.abs(getRandom() * 2 * 4))
-
-  // Corner radii
-  const rx = Math.min(w / 2, sw * 2)
-  const ry = Math.min(h / 2, sw * 2)
-
-  // Number of points per side
-  const px = Math.max(8, Math.floor(w / 16))
-  const py = Math.max(8, Math.floor(h / 16))
-
-  // Inset each line by the corner radii and let the freehand algo
-  // interpolate points for the corners.
-  const lines = Utils.rotateArray(
-    [
-      Vec.pointsBetween(Vec.add(tl, [rx, 0]), Vec.sub(tr, [rx, 0]), px),
-      Vec.pointsBetween(Vec.add(tr, [0, ry]), Vec.sub(br, [0, ry]), py),
-      Vec.pointsBetween(Vec.sub(br, [rx, 0]), Vec.add(bl, [rx, 0]), px),
-      Vec.pointsBetween(Vec.sub(bl, [0, ry]), Vec.add(tl, [0, ry]), py),
-    ],
-    rm
-  )
-
-  // For the final points, include the first half of the first line again,
-  // so that the line wraps around and avoids ending on a sharp corner.
-  // This has a bit of finesse and magic—if you change the points between
-  // function, then you'll likely need to change this one too.
-
-  const points = [...lines.flat(), ...lines[0]].slice(
-    5,
-    Math.floor((rm % 2 === 0 ? px : py) / -2) + 3
-  )
-
-  return {
-    points,
-  }
-}
-
-function getDrawStrokeInfo(shape: RectangleShape) {
-  const { points } = getRectangleDrawPoints(shape)
-
-  const { strokeWidth } = getShapeStyle(shape.style)
-
-  const options = {
-    size: strokeWidth,
-    thinning: 0.65,
-    streamline: 0.3,
-    smoothing: 1,
-    simulatePressure: false,
-    last: true,
-  }
-
-  return { points, options }
-}
-
-function getRectanglePath(shape: RectangleShape) {
-  const { points, options } = getDrawStrokeInfo(shape)
-
-  const stroke = getStroke(points, options)
-
-  return Utils.getSvgPathFromStroke(stroke)
-}
-
-function getRectangleIndicatorPathTDSnapshot(shape: RectangleShape) {
-  const { points, options } = getDrawStrokeInfo(shape)
-
-  const strokePoints = getStrokePoints(points, options)
-
-  return Utils.getSvgPathFromStroke(
-    strokePoints.map((pt) => pt.point.slice(0, 2)),
-    false
-  )
-}
-
-const Wrapper = styled('div', {
-  width: '100%',
-  height: '100%',
-})
